+++ src/stats_duck_extension.cpp
@@ -10,6 +10,7 @@
 #include "rank_correlation_function.hpp"
 #include "sign_test_function.hpp"
 #include "adjust_p_function.hpp"
+#include "poibin_function.hpp"
 #include "table_one_function.hpp"
 #include "normality_function.hpp"
 #include "ks_test_function.hpp"
@@ -51,6 +52,9 @@ static void LoadInternal(ExtensionLoader &loader) {
 
 	// Multiple-testing corrections
 	RegisterAdjustP(loader);
+
+	// Poisson Binomial CDF
+	RegisterPoibinCdf(loader);
 
 	// Table 1 helper
 	RegisterTableOne(loader);